# docs/api_contract.yaml
...
openapi: 3.1.0
info:
  title: AutorIA API
  version: 1.0.0
  description: |
    REST API for AutorIA — style-conditioned generation and Authorship Passport
    verification (EU AI Act Art. 50).

    **Owners:** P3 (backend), P1 (frontend client). Pair session closes this contract in Sprint 1.

    **Not in scope (MVP):** auth, streaming, caching, user accounts, i18n on API messages.
  contact:
    name: AutorIA Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development (FastAPI + Uvicorn)
  - url: https://api.autoria.app
    description: Production (Railway)
...
paths:
...
  /api/authors:
    get:
      tags: [authors]
      operationId: listAuthors
      summary: List authors
      description: Returns all authors (3 preloaded + any added via document upload).
      responses:
        "200":
          description: Author list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorSummary"
              examples:
                preloaded:
                  summary: Three seeded authors
                  value:
                    - id: austen
                      name: Jane Austen
                      slug: austen
                      has_style_profile: true
                      n_documents: 3
                    - id: dickens
                      name: Charles Dickens
                      slug: dickens
                      has_style_profile: true
                      n_documents: 3
                    - id: poe
                      name: Edgar Allan Poe
                      slug: poe
                      has_style_profile: true
                      n_documents: 15
        "500":
          $ref: "#/components/responses/InternalError"
  /api/authors/{author_id}/style-profile:
    get:
      tags: [authors]
      operationId: getAuthorStyleProfile
      summary: Get StyleProfile (Style DNA)
      description: |
        Returns the latest computed StyleProfile v1.0 for the author.
        Used by the Style DNA panel (radar chart + 2D map).
      parameters:
        - $ref: "#/components/parameters/AuthorId"
      responses:
        "200":
          description: StyleProfile JSON v1.0
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfile"
        "404":
          description: Author unknown or StyleProfile not yet computed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notComputed:
                  value:
                    error: not_found
                    message: StyleProfile not yet computed for this author
        "500":
          $ref: "#/components/responses/InternalError"
  /api/authors/{author_id}/style-profile/recompute:
    post:
      tags: [authors]
      operationId: recomputeAuthorStyleProfile
      summary: Trigger manual StyleProfile recompute
      description: |
        Enqueues an async StyleProfile recompute for the author.
        Returns 202 immediately; poll GET style-profile to detect completion.
        The stub pipeline (Sprint 2 P2 entrypoint not yet wired) inserts a
        placeholder-valued but schema-valid StyleProfile row.
      parameters:
        - $ref: "#/components/parameters/AuthorId"
      responses:
        "202":
          description: Recompute accepted and enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileRecomputeAccepted"
              examples:
                accepted:
                  value:
                    status: computing
                    estimated_seconds: 30
        "404":
          description: Unknown author_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/authors/{author_id}/documents:
    post:
      tags: [authors]
      operationId: uploadAuthorDocument
      summary: Upload a document to an author's corpus
      description: |
        Accepts a `.txt` or `.md` file (multipart) or inline JSON `{ title, text }`.
        Triggers async chunking, embedding, and StyleProfile recompute.
        Minimum corpus rules apply at profile compute time (see MVP §4.1).
      parameters:
        - $ref: "#/components/parameters/AuthorId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Plain-text or Markdown file (.txt, .md)
                title:
                  type: string
                  maxLength: 500
                  description: Optional display title; defaults to filename
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUploadJson"
      responses:
        "202":
          description: Document accepted for async processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadAccepted"
        "400":
          description: Invalid file type, empty content, or malformed body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown author_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    AuthorId:
      name: author_id
      in: path
      required: true
      description: Author identifier (slug), e.g. `dickens`, `austen`, `poe`
      schema:
        type: string
        pattern: ^[a-z][a-z0-9_-]{0,63}$
      examples:
        dickens:
          value: dickens
  responses:
    ValidationError:
      description: Request body failed Pydantic validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples: [not_found, bad_request, internal_error]
        message:
          type: string
          description: Human-readable English message for UI or logs
    ValidationErrorResponse:
      type: object
      required: [error, message, details]
      properties:
        error:
          type: string
          const: validation_error
        message:
          type: string
          example: Request validation failed
        details:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
    AuthorSummary:
      type: object
      required: [id, name, slug, has_style_profile, n_documents]
      properties:
        id:
          type: string
          description: Stable author identifier (same as slug for preloaded authors)
          example: dickens
        name:
          type: string
          example: Charles Dickens
        slug:
          type: string
          example: dickens
        has_style_profile:
          type: boolean
          description: True when a StyleProfile v1.0 exists and is served by GET style-profile
        n_documents:
          type: integer
          minimum: 0
          description: Number of ingested documents in the corpus
    DocumentUploadJson:
      type: object
      required: [title, text]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        text:
          type: string
          minLength: 1
          description: Raw document body (UTF-8)
    DocumentUploadAccepted:
      type: object
      required: [document_id, status]
      properties:
        document_id:
          type: string
          format: uuid
        status:
          type: string
          const: processing
          description: Async pipeline state; poll style-profile until recomputed
    StyleProfileRecomputeAccepted:
      type: object
      required: [status, estimated_seconds]
      properties:
        status:
          type: string
          const: computing
          description: Always "computing" — recompute is async
        estimated_seconds:
          type: integer
          minimum: 0
          description: Rough wall-clock estimate; derived from corpus n_tokens (max(30, n_tokens // 2000))
    StyleProfile:
      type: object
      description: Style DNA fingerprint v1.0 — see docs/MVP.md §4.2
      required:
        - schema_version
        - author_id
        - computed_at
        - corpus_stats
        - lexical
        - syntactic
        - stylistic
        - distinctive_vocab
        - semantic_centroid
        - embedding_umap_2d
      properties:
        schema_version:
          type: string
          const: "1.0"
        author_id:
          type: string
        computed_at:
          type: string
          format: date-time
        corpus_stats:
          $ref: "#/components/schemas/CorpusStats"
        lexical:
          $ref: "#/components/schemas/LexicalFeatures"
        syntactic:
          $ref: "#/components/schemas/SyntacticFeatures"
        stylistic:
          $ref: "#/components/schemas/StylisticFeatures"
        distinctive_vocab:
          type: array
          items:
            $ref: "#/components/schemas/DistinctiveTerm"
        semantic_centroid:
          type: array
          description: 768-dimensional mean embedding vector
          items:
            type: number
          minItems: 768
          maxItems: 768
        embedding_umap_2d:
          $ref: "#/components/schemas/EmbeddingUmap2d"
    CorpusStats:
      type: object
      required: [n_documents, n_tokens, n_sentences]
      properties:
        n_documents:
          type: integer
          minimum: 0
        n_tokens:
          type: integer
          minimum: 0
        n_sentences:
          type: integer
          minimum: 0
    LexicalFeatures:
      type: object
      required: [mattr_500, avg_word_length, hapax_ratio]
      properties:
        mattr_500:
          type: number
          description: Moving Average TTR (window 500)
        avg_word_length:
          type: number
        hapax_ratio:
          type: number
    SyntacticFeatures:
      type: object
      required:
        - avg_sentence_length_tokens
        - std_sentence_length_tokens
        - subordination_ratio
        - passive_voice_ratio
        - noun_to_verb_ratio
      properties:
        avg_sentence_length_tokens:
          type: number
        std_sentence_length_tokens:
          type: number
        subordination_ratio:
          type: number
        noun_to_verb_ratio:
          type: number
        passive_voice_ratio:
          type: number
    StylisticFeatures:
      type: object
      required:
        [
          punct_distribution,
          pos_distribution,
          dialogue_ratio,
          first_person_ratio,
        ]
      properties:
        punct_distribution:
          type: object
          additionalProperties:
            type: number
          description: Relative frequency per punctuation mark
        pos_distribution:
          type: object
          additionalProperties:
            type: number
          description: Relative frequency per POS tag (spaCy labels)
        dialogue_ratio:
          type: number
          description: >-
            Fraction of non-punctuation TOKENS inside straight ASCII double
            quotes, in [0, 1]. Not a fraction of sentences.
            See docs/style_features.md §3.3.
        first_person_ratio:
          type: number
          description: >-
            First-person singular pronouns per 1,000 TOKENS, >= 0 and NOT
            bounded by 1 (Poe's expected range is 18-30). Not a fraction of
            sentences. See docs/style_features.md §3.4.
    DistinctiveTerm:
      type: object
      required: [term, score]
      properties:
        term:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
          description: >-
            Log-odds-ratio vs the pooled other authors (Jeffreys prior α=0.5),
            normalized to [0, 1] within each author's run. Not a TF-IDF value.
            See docs/style_features.md §4.1.
    EmbeddingUmap2d:
      type: object
      required: [centroid, spread]
      properties:
        centroid:
          type: array
          prefixItems:
            - type: number
            - type: number
          minItems: 2
          maxItems: 2
          description: UMAP 2D coordinates for the author cluster center
        spread:
          type: number
          description: Cluster dispersion metric for viz scaling
...